+++ src/codegen/index.ts
@@ -15,19 +15,245 @@ interface Columns {
 
 interface Relation {
 	name: string;
-	target: string;
+	target: CollectionDefinition;
+	unique: boolean;
+}
+
+interface CollectionDefinition {
+	id: string;
+	name: string;
+	type: Collection['type'];
+	typeName: string;
+	columns: Columns;
+	relations: Relation[];
 }
 
 export async function generateTypes({ url, email, password }: GenerateOptions) {
 	const pb = new PocketBase(url);
 	await pb.admins.authWithPassword(email, password);
 
-	const collections: Collection[] = await pb.collections.getFullList();
+	const collections = await pb.collections.getFullList<Collection>();
+	const definitions = buildCollectionDefinitions(collections);
+
+	const definition = `/**
+ * This file was @generated using typed-pocketbase
+ */
+
+// https://pocketbase.io/docs/collections/#base-collection
+export interface BaseCollectionResponse {
+	/**
+	 * 15 characters string to store as record ID.
+	 */
+	id: string;
+	/**
+	 * Date string representation for the creation date.
+	 */
+	created: string;
+	/**
+	 * Date string representation for the creation date.
+	 */
+	updated: string;
+	/**
+	 * The collection id.
+	 */
+	collectionId: string;
+	/**
+	 * The collection name.
+	 */
+	collectionName: string;
+}
+
+// https://pocketbase.io/docs/api-records/#create-record
+export interface BaseCollectionCreate {
+	/**
+	 * 15 characters string to store as record ID.
+	 * If not set, it will be auto generated.
+	 */
+	id?: string;
+}
+
+// https://pocketbase.io/docs/api-records/#update-record
+export interface BaseCollectionUpdate {}
+
+// https://pocketbase.io/docs/collections/#auth-collection
+export interface AuthCollectionResponse extends BaseCollectionResponse {
+	/**
+	 * The username of the auth record.
+	 */
+	username: string;
+	/**
+	 * Auth record email address.
+	 */
+	email: string;
+	/**
+	 * Whether to show/hide the auth record email when fetching the record data.
+	 */
+	emailVisibility: boolean;
+	/**
+	 * Indicates whether the auth record is verified or not.
+	 */
+	verified: boolean;
+}
+
+// https://pocketbase.io/docs/api-records/#create-record
+export interface AuthCollectionCreate extends BaseCollectionCreate {
+	/**
+	 * The username of the auth record.
+	 * If not set, it will be auto generated.
+	 */
+	username?: string;
+	/**
+	 * Auth record email address.
+	 */
+	email?: string;
+	/**
+	 * Whether to show/hide the auth record email when fetching the record data.
+	 */
+	emailVisibility?: boolean;
+	/**
+	 * Auth record password.
+	 */
+	password: string;
+	/**
+	 * Auth record password confirmation.
+	 */
+	passwordConfirm: string;
+	/**
+	 * Indicates whether the auth record is verified or not.
+	 * This field can be set only by admins or auth records with "Manage" access.
+	 */
+	verified?: boolean;
+}
+
+// https://pocketbase.io/docs/api-records/#update-record
+export interface AuthCollectionUpdate {
+	/**
+	 * The username of the auth record.
+	 */
+	username?: string;
+	/**
+	 * The auth record email address.
+	 * This field can be updated only by admins or auth records with "Manage" access.
+	 * Regular accounts can update their email by calling "Request email change".
+	 */
+	email?: string;
+	/**
+	 * Whether to show/hide the auth record email when fetching the record data.
+	 */
+	emailVisibility?: boolean;
+	/**
+	 * Old auth record password.
+	 * This field is required only when changing the record password. Admins and auth records with "Manage" access can skip this field.
+	 */
+	oldPassword?: string;
+	/**
+	 * New auth record password.
+	 */
+	password?: string;
+	/**
+	 * New auth record password confirmation.
+	 */
+	passwordConfirm?: string;
+	/**
+	 * Indicates whether the auth record is verified or not.
+	 * This field can be set only by admins or auth records with "Manage" access.
+	 */
+	verified?: boolean;
+}
+
+// https://pocketbase.io/docs/collections/#view-collection
+export interface ViewCollectionRecord {
+	id: string;
+}
+
+// utilities
+
+type MaybeArray<T> = T | T[];
+
+${definitions.map(createCollectionTypes).join('\n\n')}
+
+// ===== Schema =====
+
+export type Schema = {
+	${definitions
+		.map(({ name }) => `${name}: ${pascalCase(name)}Collection;`)
+		.join(`\n\t`)}
+};`;
+
+	return definition;
+}
+
+function createCollectionTypes({
+	name,
+	relations,
+	columns,
+	type,
+	typeName
+}: CollectionDefinition) {
+	const prefix = pascalCase(type);
+	const base = `${prefix}Collection`;
+
+	let out = `// ===== ${name} =====`;
+
+	const { response, create, update } = columns;
+
+	const responseColumns = [`collectionName: '${name}';`, ...response];
+
+	out += `\n\nexport interface ${typeName}Response extends ${base}Response {\n\t${responseColumns.join(
+		`\n\t`
+	)}\n}`;
+
+	if (type !== 'view') {
+		const createBody = create.length
+			? `{\n\t${create.join(`\n\t`)}\n}`
+			: '{}';
+
+		out += `\n\nexport interface ${typeName}Create extends ${base}Create ${createBody}`;
+
+		const updateBody = update.length
+			? `{\n\t${update.join(`\n\t`)}\n}`
+			: '{}';
+
+		out += `\n\nexport interface ${typeName}Update extends ${base}Update ${updateBody}`;
+	}
+
+	const createRelations = () => {
+		return relations
+			.map(
+				(r) =>
+					`${/^\w+$/.test(r.name) ? r.name : `'${r.name}'`}: ${
+						r.target.typeName
+					}Collection${r.unique ? '' : '[]'};`
+			)
+			.join('\n\t\t');
+	};
+
+	const collectionBody = [
+		`type: '${type}';`,
+		`collectionId: string;`,
+		`collectionName: '${name}';`,
+		`response: ${typeName}Response;`,
+		type !== 'view' && `create: ${typeName}Create;`,
+		type !== 'view' && `update: ${typeName}Update;`,
+		`relations: ${
+			relations.length === 0
+				? 'Record<string, never>;'
+				: `{\n\t\t${createRelations()}\n\t};`
+		}`
+	].filter(Boolean);
+
+	out += `\n\nexport interface ${typeName}Collection {
+	${collectionBody.join('\n\t')}
+}`;
+
+	return out;
+}
+
+function buildCollectionDefinitions(collections: Collection[]) {
 	const deferred: Array<() => void> = [];
+	const definitions = new Map<string, CollectionDefinition>();
 
-	const tables = collections.map((collection) => {
-		const typeName = pascalCase(collection.name);
-
+	for (const collection of collections) {
 		const columns: Columns = {
 			create: [],
 			update: [],
@@ -35,13 +261,18 @@ export async function generateTypes({ url, email, password }: GenerateOptions) {
 		};
 		const relations: Relation[] = [];
 
-		collection.schema.forEach((field) => {
+		for (const field of collection.schema) {
 			getFieldType(field, columns);
 
 			if (field.type === 'relation') {
 				deferred.push(() => {
-					const target = tableMap.get(field.options.collectionId);
+					const from = definitions.get(collection.id);
+					const target = definitions.get(field.options.collectionId);
 
+					if (!from)
+						throw new Error(
+							`Collection ${collection.id} not found for relation ${collection.name}.${field.name}`
+						);
 					if (!target)
 						throw new Error(
 							`Collection ${field.options.collectionId} not found for relation ${collection.name}.${field.name}`
@@ -49,9 +280,8 @@ export async function generateTypes({ url, email, password }: GenerateOptions) {
 
 					relations.push({
 						name: field.name,
-						target: `${target.typeName}Collection${
-							field.options.maxSelect === 1 ? '' : '[]'
-						}`
+						target,
+						unique: field.options.maxSelect === 1
 					});
 
 					/**
@@ -62,168 +292,42 @@ export async function generateTypes({ url, email, password }: GenerateOptions) {
 					const indicies = collection.indexes.map(parseIndex);
 
 					const isUnique = indicies.some(
-						(i) =>
-							i &&
-							i.unique &&
-							i.fields.length === 1 &&
-							i.fields[0] === field.name
+						(index) =>
+							index &&
+							index.unique &&
+							index.fields.length === 1 &&
+							index.fields[0] === field.name
 					);
 
 					target.relations.push({
-						name: `'${collection.name}(${field.name})'`,
-						target: `${typeName}Collection${isUnique ? '' : '[]'}`
+						name: `${collection.name}(${field.name})`,
+						target: from,
+						unique: isUnique
 					});
 				});
 			}
-		});
+		}
 
-		return {
+		definitions.set(collection.id, {
 			id: collection.id,
 			name: collection.name,
 			type: collection.type,
-			typeName,
 			columns,
-			relations
-		};
-	});
-
-	const tableMap = new Map(tables.map((t) => [t.id, t]));
+			relations,
+			typeName: pascalCase(collection.name)
+		});
+	}
 
 	deferred.forEach((c) => c());
 
-	const indent = '\t';
-
-	const definition = `
-/**
- * This file was @generated using typed-pocketbase
- */
-
-// https://pocketbase.io/docs/collections/#base-collection
-interface BaseCollectionRecord {
-	id: string;
-	created: string;
-	updated: string;
-	collectionId: string;
-	collectionName: string;
-}
-
-// https://pocketbase.io/docs/api-records/#create-record
-interface BaseCollectionRecordCreate {
-	id?: string;
-}
-
-// https://pocketbase.io/docs/collections/#auth-collection
-interface AuthCollectionRecord extends BaseCollectionRecord {
-	username: string;
-	email: string;
-	emailVisibility: boolean;
-	verified: boolean;
-}
-
-// https://pocketbase.io/docs/api-records/#create-record
-interface AuthCollectionRecordCreate extends BaseCollectionRecordCreate {
-	username?: string;
-	email?: string;
-	emailVisibility?: boolean;
-	verified?: boolean;
-	password: string;
-	passwordConfirm: string;
-}
-
-// https://pocketbase.io/docs/api-records/#update-record
-interface AuthCollectionRecordUpdate {
-	username?: string;
-	email?: string;
-	emailVisibility?: boolean;
-	verified?: boolean;
-	password?: string;
-	passwordConfirm?: string;
-}
-
-// https://pocketbase.io/docs/collections/#view-collection
-interface ViewCollectionRecord {
-	id: string;
-}
-
-// utilities
-
-type MaybeArray<T> = T | T[];
-
-${tables
-	.map((t) => {
-		const baseRecord =
-			t.type === 'base'
-				? 'BaseCollectionRecord'
-				: t.type === 'auth'
-					? 'AuthCollectionRecord'
-					: 'ViewCollectionRecord';
-
-		return `
-// ===== ${t.name} =====
-
-export interface ${t.typeName}Response extends ${baseRecord} {
-	collectionName: '${t.name}';${!t.columns.response.length ? '' : `
-	${t.columns.response.join('\n' + indent)}`}
-}
-${
-	// view collections are readonly
-	t.type === 'view'
-		? ''
-		: `
-export interface ${t.typeName}Create extends ${t.type === "base" ? "BaseCollectionRecordCreate" : "AuthCollectionRecordCreate"} ${!t.columns.create.length ? '{}' : `{
-	${t.columns.create.join('\n' + indent)}
-}`}
-
-export interface ${t.typeName}Update${t.type === "base" ? "" : " extends AuthCollectionRecordUpdate"} ${!t.columns.update.length ? '{}' : `{
-	${t.columns.update.join('\n' + indent)}
-}`}
-`
-}
-export interface ${t.typeName}Collection {
-	type: '${t.type}';
-	collectionId: string;
-	collectionName: '${t.name}';
-	response: ${t.typeName}Response;${
-		t.type === 'view'
-			? ''
-			: `
-	create: ${t.typeName}Create;
-	update: ${t.typeName}Update;`
-	}
-	relations: ${
-		t.relations.length === 0
-			? 'Record<string, never>'
-			: `{
-		${t.relations
-			.map((col) => `${col.name}: ${col.target};`)
-			.join('\n' + indent.repeat(2))}
-	}`
-	};
-}
-
-`.trim();
-	})
-	.join('\n\n')}
-
-// ===== Schema =====
-
-export type Schema = {
-	${tables
-		.map(({ name, typeName }) => `${name}: ${typeName}Collection;`)
-		.join('\n' + indent)}
-};
-`.trim();
-
-	return definition;
+	return Array.from(definitions.values());
 }
 
 function getFieldType(field: Field, { response, create, update }: Columns) {
-	const req = field.required ? '' : '?';
-
 	const addResponse = (type: string, name = field.name) =>
 		response.push(`${name}: ${type};`);
 	const addCreate = (type: string, name = field.name) =>
-		create.push(`${name}${req}: ${type};`);
+		create.push(`${name}${field.required ? '' : '?'}: ${type};`);
 	const addUpdate = (type: string, name = field.name) =>
 		update.push(`${name}?: ${type};`);
 	const addAll = (type: string) => {
@@ -264,12 +368,11 @@ function getFieldType(field: Field, { response, create, update }: Columns) {
 		}
 		case 'select': {
 			const single = field.options.maxSelect === 1;
-      const values = !field.required && single
-				? ["", ...field.options.values]
-				: field.options.values;
-			const singleType = values
-				.map((v) => `'${v}'`)
-				.join(' | ');
+			const values =
+				!field.required && single
+					? ['', ...field.options.values]
+					: field.options.values;
+			const singleType = values.map((v) => `'${v}'`).join(' | ');
 			const type = single ? `${singleType}` : `MaybeArray<${singleType}>`;
 
 			addResponse(single ? singleType : `Array<${singleType}>`);
@@ -285,7 +388,7 @@ function getFieldType(field: Field, { response, create, update }: Columns) {
 		case 'relation': {
 			const singleType = 'string';
 			const single = field.options.maxSelect === 1;
-			const type = single ? `${singleType}` : `MaybeArray<${singleType}>`;
+			const type = single ? singleType : `MaybeArray<${singleType}>`;
 
 			addResponse(single ? singleType : `Array<${singleType}>`);
 			addCreate(type);
@@ -297,15 +400,13 @@ function getFieldType(field: Field, { response, create, update }: Columns) {
 			break;
 		}
 		case 'file': {
-			const singleType = 'string';
 			const single = field.options.maxSelect === 1;
-			const type = single ? `${singleType}` : `MaybeArray<${singleType}>`;
 
-			addResponse(single ? singleType : `Array<${singleType}>`);
-			addCreate(type);
-			addUpdate(type);
+			addResponse(single ? 'string' : `Array<string>`);
+			addCreate(single ? `File | null` : `MaybeArray<File>`);
+			addUpdate(single ? `File | null` : `MaybeArray<File>`);
 			if (!single) {
-				addUpdate(type, `'${field.name}-'`);
+				addUpdate('string', `'${field.name}-'`);
 			}
 			break;
 		}