+++ src/filter.ts
@@ -1,8 +1,4 @@
-import {
-	BaseRecord,
-	GenericCollection,
-	RecordWithExpandToDotPath
-} from './types.js';
+import type { BaseRecord } from './types.js';
 
 type ActualFilter<T extends any, K extends keyof T = keyof T> = [
 	K,
@@ -30,18 +26,14 @@ export type FilterOperand =
 
 export type FilterParam<T extends BaseRecord> = { __record__?: T } & string;
 
-export type Filter<T extends GenericCollection> = FilterParam<
-	RecordWithExpandToDotPath<T>
->;
-
-export type FilterInput<T extends BaseRecord> =
+export type Filter<T extends BaseRecord> =
 	| ActualFilter<T>
 	| FilterParam<T>
 	| false
 	| null
 	| undefined;
 
-function serializeFilter([key, op, val]: ActualFilter<any>) {
+export function serializeFilterTuple([key, op, val]: ActualFilter<any>) {
 	const type = typeof val;
 	if (type === 'boolean' || type === 'number') {
 		val = val.toString();
@@ -55,19 +47,20 @@ function serializeFilter([key, op, val]: ActualFilter<any>) {
 		val = "'" + JSON.stringify(val).replace(/'/g, "\\'") + "'";
 	}
 
-	return `${String(key)}${op}${val}`;
+	return `${String(key)} ${op} ${val}`;
 }
 
-export function serializeFilters(filters: FilterInput<any>[]) {
-	return filters
-		.filter(Boolean)
-		.map((filter) =>
-			Array.isArray(filter) ? serializeFilter(filter) : filter
-		);
+export function serializeFilter(filter: Filter<any>): string | null {
+	if (!filter) return null;
+	return Array.isArray(filter) ? serializeFilterTuple(filter) : filter;
+}
+
+export function serializeFilters(filters: Filter<any>[]) {
+	return filters.filter((val) => !!val).map(serializeFilter);
 }
 
 export function and<T extends BaseRecord>(
-	...filters: FilterInput<T>[]
+	...filters: Filter<T>[]
 ): FilterParam<T> {
 	const str = serializeFilters(filters).join(' && ');
 	if (!str.length) return '';
@@ -75,7 +68,7 @@ export function and<T extends BaseRecord>(
 }
 
 export function or<T extends BaseRecord>(
-	...filters: FilterInput<T>[]
+	...filters: Filter<T>[]
 ): FilterParam<T> {
 	const str = serializeFilters(filters).join(' || ');
 	if (!str.length) return '';
@@ -86,54 +79,54 @@ export function eq<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '=', value]);
+	return serializeFilterTuple([column, '=', value]);
 }
 
 export function neq<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '!=', value]);
+	return serializeFilterTuple([column, '!=', value]);
 }
 
 export function gt<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '>', value]);
+	return serializeFilterTuple([column, '>', value]);
 }
 
 export function gte<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '>=', value]);
+	return serializeFilterTuple([column, '>=', value]);
 }
 
 export function lt<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '<', value]);
+	return serializeFilterTuple([column, '<', value]);
 }
 
 export function lte<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '<=', value]);
+	return serializeFilterTuple([column, '<=', value]);
 }
 
 export function like<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '~', value]);
+	return serializeFilterTuple([column, '~', value]);
 }
 
 export function nlike<T extends BaseRecord, Key extends keyof T>(
 	column: Key,
 	value: T[Key]
 ): FilterParam<T> {
-	return serializeFilter([column, '!~', value]);
+	return serializeFilterTuple([column, '!~', value]);
 }